Create ExtractMthEnd
extract month end data 
If the DTD value of final trading date in one month is NaN, it cannot be set as the month-end data and the value from the previous trading date should be used instead. 
If all the DTD value in a single month are NaN, we should set the last trading date value as the month-end value.